Describe parameter support
Changed how the types.json value is emitted by the extension. It is no
longer required for the server to launch.  If the describe
(`-d true`,`--describe true`,`describe true`) argument is sent the
application will output the type spec to stdout and exit.  This is to be used
by Bicep during extension publishing,`bicep publish-extension`, to generate the package, avoiding
the need to launch the server.